refactor: Usability and Consumability Improvements; ci: pre-commit

+++ horde_model_reference/legacy/classes/legacy_converters.py
@@ -1,3 +1,4 @@
+"""The classes used to convert the legacy model reference to the new format."""
 import glob
 import json
 import typing
@@ -7,33 +8,39 @@ from pathlib import Path
 from pydantic import ValidationError
 from typing_extensions import override
 
-from horde_model_reference import path_consts
+from horde_model_reference import (
+    BASE_PATH,
+    DEFAULT_SHOWCASE_FOLDER_NAME,
+    LEGACY_REFERENCE_FOLDER,
+    LOG_FOLDER,
+    MODEL_PURPOSE_LOOKUP,
+    MODEL_REFERENCE_CATEGORIES,
+    path_consts,
+)
 from horde_model_reference.legacy.classes.staging_model_database_records import (
     MODEL_REFERENCE_LEGACY_TYPE_LOOKUP,
     Legacy_Generic_ModelReference,
     Legacy_StableDiffusion_ModelRecord,
-    Legacy_StableDiffusion_ModelReference,
+    Staging_StableDiffusion_ModelReference,
     StagingLegacy_Config_DownloadRecord,
     StagingLegacy_Config_FileRecord,
     StagingLegacy_Generic_ModelRecord,
 )
-from horde_model_reference.meta_consts import MODEL_PURPOSE_LOOKUP
 from horde_model_reference.model_reference_records import (
     MODEL_REFERENCE_TYPE_LOOKUP,
     StableDiffusion_ModelReference,
 )
 from horde_model_reference.path_consts import (
-    BASE_PATH,
-    DEFAULT_SHOWCASE_FOLDER_NAME,
     GITHUB_REPO_URL,
-    LEGACY_REFERENCE_FOLDER,
-    MODEL_REFERENCE_CATEGORIES,
     PACKAGE_NAME,
 )
 from horde_model_reference.util import model_name_to_showcase_folder_name
 
 
 class BaseLegacyConverter:
+    """The logic applicable to all legacy model reference converters.
+    See normalize_and_convert() for the order of operations critical to the conversion process."""
+
     legacy_folder_path: Path
     """The folder path to the legacy model reference."""
     legacy_database_path: Path
@@ -59,14 +66,22 @@ class BaseLegacyConverter:
     print_errors: bool = True
     """Whether to print errors in the conversion to `stdout`."""
 
+    log_folder: Path = LOG_FOLDER
+    """The folder to write the validation error log to."""
+
+    dry_run: bool = False
+    """If true, don't write out the converted database or any log files."""
+
     def __init__(
         self,
         *,
         legacy_folder_path: str | Path = LEGACY_REFERENCE_FOLDER,
         target_file_folder: str | Path = BASE_PATH,
+        log_folder: str | Path = LOG_FOLDER,
         model_reference_category: MODEL_REFERENCE_CATEGORIES,
         print_errors: bool = True,
         debug_mode: bool = False,
+        dry_run: bool = False,
     ):
         """Initialize an instance of the LegacyConverterBase class.
 
@@ -76,6 +91,7 @@ class BaseLegacyConverter:
             model_reference_category (MODEL_REFERENCE_CATEGORIES): The category of model reference to convert.
             print_errors (bool, optional): Whether to print errors in the conversion to `stdout`. Defaults to True.
             debug_mode (bool, optional): If true, include extra information in the error log. Defaults to False.
+            dry_run (bool, optional): If true, don't write out the converted database or any logs. Defaults to False.
         """
         self.all_model_records = {}
         self.all_validation_errors_log = {}
@@ -96,11 +112,15 @@ class BaseLegacyConverter:
         self.debug_mode = debug_mode
         self.print_errors = print_errors
 
+        self.log_folder = Path(log_folder)
+
+        self.dry_run = dry_run
+
     def normalize_and_convert(self) -> bool:
         """Normalizes and converts the legacy model reference database to the new format.
 
         Returns:
-            bool: True if the conversion was successful, False otherwise.
+            bool: `True` if the conversion was successful, False otherwise.
         """
         self.pre_parse_records()
         all_model_iterator = self._iterate_over_input_records(self.model_reference_type)
@@ -248,7 +268,7 @@ class BaseLegacyConverter:
         """Override and call super().parse_record(..) to perform any model category specific parsing."""
 
     def post_parse_records(self) -> None:
-        """Perform any post parsing tasks."""
+        """Override and call super().post_parse_records() to perform any model category specific post parsing."""
         for model_record in self.all_model_records.values():
             model_record.model_purpose = MODEL_PURPOSE_LOOKUP[self.model_reference_category]
         pass
@@ -264,6 +284,9 @@ class BaseLegacyConverter:
         except ValidationError as e:
             print(f"CRITICAL: Failed to convert to new model reference type {type_to_convert_to}.")
             raise e
+
+        if self.dry_run:
+            return
 
         with open(self.converted_database_file_path, "w") as new_model_reference_file:
             new_model_reference_file.write(
@@ -290,7 +313,10 @@ class BaseLegacyConverter:
 
     def write_out_validation_errors(self) -> None:
         """Write out the validation errors."""
-        log_file = self.converted_folder_path.joinpath(self.model_reference_category + ".log")
+        if self.dry_run:
+            return
+
+        log_file = self.log_folder.joinpath(self.model_reference_category + ".log")
         with open(log_file, "w") as validation_errors_log_file:
             validation_errors_log_file.write(
                 json.dumps(
@@ -491,7 +517,7 @@ class LegacyStableDiffusionConverter(BaseLegacyConverter):
         if len(sanity_check) != len(self.all_model_records):
             raise ValueError("CRITICAL: Not all records are of the correct type.")
 
-        modelReference = Legacy_StableDiffusion_ModelReference(
+        modelReference = Staging_StableDiffusion_ModelReference(
             baseline=self.all_baseline_categories,
             styles=self.all_styles,
             tags=self.all_tags,
@@ -509,6 +535,9 @@ class LegacyStableDiffusionConverter(BaseLegacyConverter):
             print(e)
             print("CRITICAL: Failed to convert to new model reference type.")
             raise e
+
+        if self.dry_run:
+            return
 
         with open(self.converted_database_file_path, "w") as testfile:
             testfile.write(json.dumps(models_in_doc_root, indent=4))
